+++ Source/Scene/GlobeSurfaceTileProvider.js
@@ -4,6 +4,7 @@ define([
         '../Core/Cartesian2',
         '../Core/Cartesian3',
         '../Core/Cartesian4',
+        '../Core/Color',
         '../Core/defined',
         '../Core/defineProperties',
         '../Core/destroyObject',
@@ -34,6 +35,7 @@ define([
         Cartesian2,
         Cartesian3,
         Cartesian4,
+        Color,
         defined,
         defineProperties,
         destroyObject,
@@ -118,9 +120,33 @@ define([
             wireframe : false,
             boundingSphereTile : undefined
         };
+
+        this._baseColor = undefined;
+        this._firstPassInitialColor = undefined;
+        this.baseColor = new Color(0.0, 0.0, 0.5, 1.0);
     };
 
     defineProperties(GlobeSurfaceTileProvider.prototype, {
+        /**
+         * Gets or sets the color of the globe when no imagery is available.
+         * @memberof GlobeSurfaceTileProvider.prototype
+         * @type {Color}
+         */
+        baseColor : {
+            get : function() {
+                return this._baseColor;
+            },
+            set : function(value) {
+                //>>includeStart('debug', pragmas.debug);
+                if (!defined(value)) {
+                    throw new DeveloperError('value is required.');
+                }
+                //>>includeEnd('debug');
+
+                this._baseColor = value;
+                this._firstPassInitialColor = Cartesian4.fromColor(value, this._firstPassInitialColor);
+            }
+        },
         /**
          * Gets or sets the {@link QuadtreePrimitive} for which this provider is
          * providing tiles.  This property may be undefined if the provider is not yet associated
@@ -766,7 +792,6 @@ define([
         return context.createVertexArray(vertexArray._attributes, wireframeIndexBuffer);
     }
 
-    var firstPassInitialColor = new Cartesian4(0.0, 0.0, 0.5, 1.0);
     var otherPassesInitialColor = new Cartesian4(0.0, 0.0, 0.0, 0.0);
 
     function addDrawCommandsForTile(tileProvider, tile, context, frameState, commandList) {
@@ -848,7 +873,7 @@ define([
         var otherPassesRenderState = tileProvider._blendRenderState;
         var renderState = firstPassRenderState;
 
-        var initialColor = firstPassInitialColor;
+        var initialColor = tileProvider._firstPassInitialColor;
 
         do {
             var numberOfDayTextures = 0;