+++ CHANGES.md
@@ -15,6 +15,8 @@ a url to a glTF file.  [#2078](https://github.com/AnalyticalGraphicsInc/cesium/i
 * Clamp of `Camera.tilt` to [-pi / 2, pi / 2] instead of [0, pi / 2].
 * Add `Camera.roll`.
 * Added `Globe.showWaterEffect` to enable/disable the water effect for supported terrain providers.
+* Added `Globe.baseColor` to set the color of the globe when no imagery is available.
+* Fixed a bug that could crash a crash when quickly adding and removing imagery layers.
 
 ### 1.1 - 2014-09-02